requirements.txt
# Backend
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.30.0
httpx==0.27.0
python-dotenv==1.0.1

# AI / ML
scikit-learn==1.4.2
xgboost==2.0.3
stable-baselines3==2.3.2
gymnasium==0.29.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
matplotlib==3.9.0

# Integrations
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
twilio==9.1.0

# Testing
pytest==8.2.0
httpx==0.27.0
